DSND-3213: Get cucumber tests running with make command
* Upgraded cucumber tests to use JUnit 5.
* Change itest classes to have IT suffix instead of ITest.
* The cucumber upgrade seemed to fix the problem - still don't know what was causing the issue with the tests not running with a make command.
