--- src/itest/java/uk/gov/companieshouse/pscdataapi/config/AbstractMongoConfig.java
@@ -1,22 +1,0 @@
-package uk.gov.companieshouse.pscdataapi.config;
-
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.MongoDBContainer;
-import org.testcontainers.utility.DockerImageName;
-
-/**
- * Mongodb configuration runs on test container.
- */
-public class AbstractMongoConfig {
-
-    public static final MongoDBContainer mongoDBContainer = new MongoDBContainer(
-            DockerImageName.parse("mongo:4"));
-
-    @DynamicPropertySource
-    public static void setProperties(DynamicPropertyRegistry registry) {
-        registry.add("feature.psc_individual_full_record_get", () -> true);
-        registry.add("spring.data.mongodb.uri", mongoDBContainer::getReplicaSetUrl);
-        mongoDBContainer.start();
-    }
-}
